+++ src/app/[variants]/(main)/chat/components/topic/features/Topic/TopicListContent/TopicItem/TopicContent.tsx
@@ -246,6 +246,13 @@ const TopicContent = memo<TopicContentProps>(({ id, title, fav, showMore }) => {
       ) : (
         <EditableText
           editing={editing}
+          onBlur={(e) => {
+            const v = (e.target as HTMLInputElement).value;
+            if (title !== v) {
+              updateTopicTitle(id, v);
+            }
+            toggleEditing(false);
+          }}
           onChangeEnd={(v) => {
             if (title !== v) {
               updateTopicTitle(id, v);