+++ src/app/[variants]/(main)/chat/components/topic/features/Topic/TopicListContent/TopicItem/index.tsx
@@ -3,7 +3,6 @@ import { createStyles } from 'antd-style';
 import qs from 'query-string';
 import { Suspense, memo, useState } from 'react';
 import { Flexbox } from 'react-layout-kit';
-import { Link } from 'react-router-dom';
 
 import { useChatStore } from '@/store/chat';
 import { useGlobalStore } from '@/store/global';
@@ -54,47 +53,45 @@ export interface ConfigCellProps {
 const TopicItem = memo<ConfigCellProps>(({ title, active, id, fav, threadId }) => {
   const { styles, cx } = useStyles();
   const toggleConfig = useGlobalStore((s) => s.toggleMobileTopic);
-  const [toggleTopic] = useChatStore((s) => [s.switchTopic]);
+  const [toggleTopic, editing] = useChatStore((s) => [s.switchTopic, s.topicRenamingId === id]);
   const activeId = useSessionStore((s) => s.activeId);
   const [isHover, setHovering] = useState(false);
 
-  const topicUrl = qs.stringifyUrl({
-    query: id ? { session: activeId, topic: id } : { session: activeId },
-    url: '/chat',
-  });
-
   return (
     <Flexbox style={{ position: 'relative' }}>
-      <Link
+      <Flexbox
+        align={'center'}
+        className={cx(styles.container, 'topic-item', active && !threadId && styles.active)}
+        distribution={'space-between'}
+        horizontal
         onClick={(e) => {
-          if (e.button === 0 && (e.metaKey || e.ctrlKey)) {
+          // 重命名时不切换话题
+          if (editing) return;
+          // Ctrl/Cmd+点击在新窗口打开
+          if (e.button === 0 && (e.metaKey || e.ctrlKey) && id) {
+            const topicUrl = qs.stringifyUrl({
+              query: { session: activeId, topic: id },
+              url: '/chat',
+            });
+            window.open(topicUrl, '_blank');
             return;
           }
-          e.preventDefault();
           toggleTopic(id);
           toggleConfig(false);
         }}
-        to={topicUrl}
+        onMouseEnter={() => {
+          setHovering(true);
+        }}
+        onMouseLeave={() => {
+          setHovering(false);
+        }}
       >
-        <Flexbox
-          align={'center'}
-          className={cx(styles.container, 'topic-item', active && !threadId && styles.active)}
-          distribution={'space-between'}
-          horizontal
-          onMouseEnter={() => {
-            setHovering(true);
-          }}
-          onMouseLeave={() => {
-            setHovering(false);
-          }}
-        >
-          {!id ? (
-            <DefaultContent />
-          ) : (
-            <TopicContent fav={fav} id={id} showMore={isHover} title={title} />
-          )}
-        </Flexbox>
-      </Link>
+        {!id ? (
+          <DefaultContent />
+        ) : (
+          <TopicContent fav={fav} id={id} showMore={isHover} title={title} />
+        )}
+      </Flexbox>
       {active && (
         <Suspense
           fallback={