+++ src/features/AgentSetting/AgentPlugin/index.tsx
@@ -4,7 +4,7 @@ import { Avatar, Button, Form, type FormGroupItemType, Tag, Tooltip } from '@lob
 import { Empty, Space, Switch } from 'antd';
 import isEqual from 'fast-deep-equal';
 import { LucideTrash2, Store } from 'lucide-react';
-import Link from 'next/link';
+import { Link, useNavigate } from 'react-router-dom';
 import { memo, useState } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 import { Center, Flexbox } from 'react-layout-kit';
@@ -28,6 +28,8 @@ const AgentPlugin = memo(() => {
   const { t } = useTranslation('setting');
 
   const [showStore, setShowStore] = useState(false);
+
+  const navigate = useNavigate();
 
   const [userEnabledPlugins, toggleAgentPlugin] = useStore((s) => [
     s.config.plugins || [],
@@ -134,11 +136,13 @@ const AgentPlugin = memo(() => {
           <Trans i18nKey={'plugin.empty'} ns={'setting'}>
             暂无安装插件，
             <Link
-              href={'/'}
               onClick={(e) => {
+                e.stopPropagation();
                 e.preventDefault();
                 setShowStore(true);
+                navigate('/discover/mcp');
               }}
+              to={'/discover/mcp'}
             >
               前往插件市场
             </Link>